+++ server/src/TickTick.Api/Dtos/Speeches/SpeechCreateDto.cs
@@ -4,6 +4,6 @@ public record SpeechCreateDto
 {
     public string Title { get; set; }
     public TimeSpan? Duration { get; set; }
-    public uint SequenceNumber { get; set; }
+    public double SequenceNumber { get; set; }
     public Guid PlaylistId { get; set; }
 }
